[HUDI-7713] Reduce code duplication
The PR does very straight-forward change to extract common part of 2
child classes out to a parent class.
